requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy<2
matplotlib>=3.7.0
pyyaml>=6.0
torch>=2.0.0
torchvision>=0.15.0
kagglehub>=0.1.0
ipython>=8.0.0
